move methods to individuals services
