+++ Services/src/main/java/com/library/exception/BookNotAvailableException.java
@@ -1,4 +1,4 @@
-package com.library.error;
+package com.library.exception;
 
 public class BookNotAvailableException extends RuntimeException{
     public BookNotAvailableException(String message){
