+++ Services/src/main/java/com/library/exception/BorrowerExistsException.java
@@ -1,4 +1,4 @@
-package com.library.error;
+package com.library.exception;
 
 public class BorrowerExistsException extends RuntimeException {
     public BorrowerExistsException(String msg) {
