+++ Services/src/main/java/com/library/exception/BorrowerThresholdException.java
@@ -1,4 +1,4 @@
-package com.library.error;
+package com.library.exception;
 
 public class BorrowerThresholdException extends RuntimeException{
     public BorrowerThresholdException(String message){
