+++ Services/src/main/java/com/library/exception/NoSuchBookException.java
@@ -1,4 +1,4 @@
-package com.library.error;
+package com.library.exception;
 
 public class NoSuchBookException extends RuntimeException {
     public NoSuchBookException(String message){
