+++ Services/src/main/java/com/library/exception/NoSuchBookLoanException.java
@@ -1,4 +1,4 @@
-package com.library.error;
+package com.library.exception;
 
 public class NoSuchBookLoanException extends RuntimeException {
     public NoSuchBookLoanException(String msg) {
