+++ Services/src/main/java/com/library/exception/NoSuchBorrowerException.java
@@ -1,4 +1,4 @@
-package com.library.error;
+package com.library.exception;
 
 public class NoSuchBorrowerException extends RuntimeException{
     public NoSuchBorrowerException(String message){
