+++ Services/src/main/java/com/library/services/FineService.java
@@ -1,0 +1,14 @@
+package com.library.services;
+
+import com.library.dto.FineDTO;
+import com.library.dto.ResponseDTO;
+
+import java.util.List;
+
+public interface FineService {
+
+    ResponseDTO calculateFines();
+    ResponseDTO payFine(int cardId);
+    List<FineDTO> getAllFines();
+    List<FineDTO> getFineForCardId(int cardId);
+}
