+++ Services/src/main/java/com/library/services/SearchService.java
@@ -1,0 +1,10 @@
+package com.library.services;
+
+import com.library.dto.BookPaginatedDTO;
+import org.springframework.data.domain.Pageable;
+
+public interface SearchService {
+    BookPaginatedDTO searchBooks(String query, Pageable pageable);
+
+    BookPaginatedDTO searchBooksForBorrower(String borrowerName, int cardId, String isbn, Pageable pageable);
+}
